+++ src/main/java/racingcar/io/constant/ErrorMessage.java
@@ -1,4 +1,4 @@
-package racingcar.io;
+package racingcar.io.constant;
 
 public final class ErrorMessage {
